+++ london-marathon/fetch_details_urls.py
@@ -29,9 +29,9 @@ def create_args():
         Parser that handles cmd arguments.
     """
 
-    parser = argparse.ArgumentParser(usage="-y <years to fetch> -o <path to output file>")
+    parser = argparse.ArgumentParser(usage="-y <years to fetch (optional)> -f <path to data file>")
     parser.add_argument("-y", dest="years", help="e.g '2017', '2014-2017', '2014,2016,2017'", required=False)
-    parser.add_argument("-f", dest="file_path", help="e.g /home/awesome/data/out.csv", required=True)
+    parser.add_argument("-f", dest="file_path", help="e.g /home/awesome/data/data.csv", required=True)
     return parser
 
 
@@ -93,10 +93,11 @@ def main():
             bot.write_dict_to_csv(urls)
         else:
             urls = StreamsBot(file_path).read_results_url_from_csv()  # parse file to get urls of results
-            details = LondonMarathonBot.get_performance_details(urls)  # get details of performances
-            out_path = os.path.join(os.path.dirname(file_path),
-                                    "out-" + str(int(time.time())) + ".csv")  # unique out file
-            StreamsBot(out_path).write_dicts_to_csv(details)  # save to output file
+            out_path = os.path.join(os.path.dirname(file_path), "out-" + str(int(time.time())) + ".csv")
+            LondonMarathonBot.async_download_performance_details(
+                urls,
+                out_path
+            )  # get details of performances
     else:
         print("Error while parsing args.")