+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## 0.1 - 2017-04-25
+
+### Tested
+- async workers for london marathon
+- refactored utils for london marathon
+
 ## 0.0 - 2017-04-25
 
 ### Released